requirements.txt
gitpython>=3.1.7,<4
pyyaml>=5.3.1,<6
Jinja2>=2.11.2,<3
